backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
boto3==1.38.3
botocore==1.38.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
fastapi==0.115.12
filelock==3.18.0
h11==0.16.0
idna==3.10
jmespath==1.0.1
packaging==25.0
platformdirs==4.3.7
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.14.1
sortedcontainers==2.4.0
starlette==0.46.2
tomlkit==0.13.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
